services/media-generator/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML Providers
openai==1.12.0
elevenlabs==0.2.27
replicate>=0.25.0

# Image Processing
pillow==10.2.0

# Video Processing
moviepy==1.0.3

# Cloud Storage
cloudinary==1.38.0
boto3==1.34.34

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
tenacity==8.2.3

# Logging
structlog==24.1.0

# Chart/Diagram generation (optional, for advanced charts)
matplotlib>=3.8.0
plotly>=5.18.0
kaleido>=0.2.1

# SVG processing (optional, for SVG diagram export)
cairosvg>=2.7.0

# Architecture Diagrams - Cloud/Infrastructure diagrams with real service icons
diagrams>=0.23.4

# Graphviz Python bindings (lightweight, no compilation needed)
graphviz>=0.20.1

# AI-based background removal for avatars (CPU mode)
# Note: Using onnxruntime instead of torch for lighter weight
rembg[cpu]>=2.0.50

# Image/Video processing utilities
imageio>=2.33.0
imageio-ffmpeg>=0.4.9
scikit-image>=0.22.0
